requirements.txt
discord.py
python-dotenv
aiohttp
puter
edge-tts
PyNaCl
